# Speed up SHAKE by using f1600x4 directly instead of shake256x4