The generic shake256x4 has considerable overhead over the underlying
permutation as it has to tranpose all incoming and outgoing data; and
work for different input and output lengths.

For thashx4 and prf_addrx4, the input size is known, the size
is smaller than the complete state and is for some quadwords identical
for all lanes.  By constructing and reading from the fourway state by
hand, the compiler can do a way better job.

This yield a nice speedup.

Old:

```
Parameters: n = 32, h = 64, d = 8, b = 14, k = 22, w = 16
Running 10 iterations.
Generating keypair.. avg.   124185.80 us (0.12 sec); median    353,740,630 cycles,      1x:    353,740,630 cycles
  - WOTS pk gen..    avg.      469.60 us (0.00 sec); median      1,297,772 cycles,    256x:    332,229,632 cycles
Signing..            avg.  1412992.80 us (1.41 sec); median  3,961,513,594 cycles,      1x:  3,961,513,594 cycles
  - FORS signing..   avg.   456295.10 us (0.46 sec); median  1,282,159,337 cycles,      1x:  1,282,159,337 cycles
  - WOTS signing..   avg.       93.50 us (0.00 sec); median        260,994 cycles,      8x:      2,087,952 cycles
  - WOTS pk gen..    avg.      469.50 us (0.00 sec); median      1,297,817 cycles,   2048x:  2,657,929,216 cycles
Verifying..          avg.     2272.40 us (0.00 sec); median      6,378,965 cycles,      1x:      6,378,965 cycles
```

New:
```
Parameters: n = 32, h = 64, d = 8, b = 14, k = 22, w = 16
Running 10 iterations.
Generating keypair.. avg.    64447.90 us (0.06 sec); median    180,103,023 cycles,      1x:    180,103,023 cycles
  - WOTS pk gen..    avg.      248.40 us (0.00 sec); median        673,056 cycles,    256x:    172,302,336 cycles
Signing..            avg.   864298.30 us (0.86 sec); median  2,430,533,372 cycles,      1x:  2,430,533,372 cycles
  - FORS signing..   avg.   333202.90 us (0.33 sec); median    934,939,607 cycles,      1x:    934,939,607 cycles
  - WOTS signing..   avg.       43.00 us (0.00 sec); median        112,891 cycles,      8x:        903,128 cycles
  - WOTS pk gen..    avg.      259.40 us (0.00 sec); median        730,774 cycles,   2048x:  1,496,625,152 cycles
Verifying..          avg.     1376.70 us (0.00 sec); median      3,761,284 cycles,      1x:      3,761,284 cycles
```

Note: the duplicated swap32 helper and `KeccakP1600times4_PermuteAll_24rounds` external declaration should be cleaned up, but I'll leave that to you.